refactored logic to fetch list of tasks so that it uses less hooks
As an alternative to using the useMemo and useCallback hooks
in the component calling a custom hook, we can write our code in such
a way that we don't pass any dependencies to the custom hook function
itself, but to a function that is returned from the custom hook. We also
move any dependencies into the useEffect callback argument function so
that we don't need to list them as dependencies to useEffect.